cts/requirements.txt
PyYAML==6.0.2
requests==2.32.3
jsonschema==4.23.0
pynacl==1.5.0
